Fixed negative event numbers. Commented out the cleanJets process since we don't want to worry about loose jet id for the 7_4_X synch exercise.

+++ UFDiMuonsAnalyzer/interface/DataFormats.h
@@ -2,7 +2,7 @@
 typedef struct {
   int run;
   int lumi;
-  int event;
+  long long int event;
   int bx;
   int orbit;
 } _EventInfo;
